Create gcdK.py
This finding the number of subarrays with gcd K. To find the number of subarrays with gcd K. we need to assume the first number of the subarray as current gcd and expanding till the end of the array. while expanding the if the current gcd of the subarray is K , increment the count of the subarray by one.